修改

+++ form_init/public/js/main.js
@@ -24,15 +24,27 @@ const vm = new Vue({
         this.newTitle = '';
       },
       removeData: function (id) {
-        let _this = this;
-        this.$http.delete(this.apiRequestUrl + '/' + id).then(function () {
-          _this.$http.get(_this.apiRequestUrl).then(function (res) {
-            this.lists = res.data;
-            let len = res.data.length;
-            let dataId = res.data[Object.keys(res.data)[Object.keys(res.data).length - 1]].id;
-            this.id = dataId ? parseInt(dataId) + 1 : 0;
-          })
-        });
+        swal({
+          title            : '确定要删除吗？',
+          type             : 'warning',
+          showCancelButton : true,
+          confirmButtonText: '确定',
+          cancelButtonText : '取消'
+        }).then((result) => {
+          if (result.value) {
+            let _this = this;
+            this.$http.delete(this.apiRequestUrl + '/' + id).then(function () {
+              _this.$http.get(_this.apiRequestUrl).then(function (res) {
+                this.lists = res.data;
+                let len = res.data.length;
+                let dataId = res.data[Object.keys(res.data)[Object.keys(res.data).length - 1]].id;
+                this.id = dataId ? parseInt(dataId) + 1 : 0;
+              })
+            }).then(function () {
+              swal('删除成功！')
+            });
+          }
+        })
       }
       ,
       editData  : function (id) {
@@ -42,40 +54,55 @@ const vm = new Vue({
       ,
       saveData  : function () {
         let _this = this;
-        let displayData = {
-          "id"    : this.id,
-          "title" : this.newTitle,
-          "date"  : new Date().toLocaleDateString(),
-          "action": "删除数据"
-        };
-        this.lists.push(displayData);
-        this.id++;
+        if (this.newTitle) {
+          let displayData = {
+            "id"    : this.id,
+            "title" : this.newTitle,
+            "date"  : new Date().toLocaleDateString(),
+            "action": "删除数据"
+          };
+          this.lists.push(displayData);
+          this.id++;
+          this.isDisplay = false;
+          this.$http.post(this.apiRequestUrl, displayData).then(function () {
+            _this.$http.get(_this.apiRequestUrl).then(function (res) {
+              _this.lists = res.data;
+              let len = res.data.length;
+              let dataId = res.data[Object.keys(res.data)[Object.keys(res.data).length - 1]].id;
+              _this.id = dataId ? parseInt(dataId) + 1 : 0;
+            })
+          });
+        } else {
+          swal('Title不能为空！');
+        }
+      },
+      cancelData: function () {
         this.isDisplay = false;
-        this.$http.post(this.apiRequestUrl, displayData).then(function () {
-          _this.$http.get(_this.apiRequestUrl).then(function (res) {
-            _this.lists = res.data;
-            let len = res.data.length;
-            let dataId = res.data[Object.keys(res.data)[Object.keys(res.data).length - 1]].id;
-            _this.id = dataId ? parseInt(dataId) + 1 : 0;
-          })
-        });
+        this.newTitle = '';
       },
       sureEdit  : function () {
         let _this = this;
-        let changeTitle = {
-          title: this.changeTitle
-        };
+        if (this.changeTitle) {
+          let changeTitle = {
+            title: this.changeTitle
+          };
+          this.isEdit = false;
+          this.$http.put(this.apiRequestUrl + '/' + this.editId, changeTitle).then(function () {
+            _this.$http.get(_this.apiRequestUrl).then(function (res) {
+              _this.lists = res.data;
+              let len = res.data.length;
+              let dataId = res.data[Object.keys(res.data)[Object.keys(res.data).length - 1]].id;
+              _this.id = dataId ? parseInt(dataId) + 1 : 0;
+            })
+          });
+        } else {
+          swal('Title不能为空！');
+        }
+      },
+      cancelEdit: function () {
         this.isEdit = false;
-        this.$http.put(this.apiRequestUrl + '/' + this.editId, changeTitle).then(function () {
-          _this.$http.get(_this.apiRequestUrl).then(function (res) {
-            _this.lists = res.data;
-            let len = res.data.length;
-            let dataId = res.data[Object.keys(res.data)[Object.keys(res.data).length - 1]].id;
-            _this.id = dataId ? parseInt(dataId) + 1 : 0;
-          })
-        });
+        this.changeTitle = '';
       }
-
     }
 
   })
